Changed simple::String hash function to return uint64_t instead of size_t

+++ simple/headers/simple_string.hpp
@@ -315,9 +315,9 @@ namespace simple {
 
 		struct Hash {
 			inline size_t operator()(const String& string) const {
-				size_t h = 37;
+				uint64_t h = 37;
 				for (char c : string) {
-					h = (h * 54059) ^ ((size_t)c * 76963);
+					h = (h * 54059) ^ ((uint64_t)c * 76963);
 				}
 				return h;
 			}
